# inline invoke