This is my attempt to improve issue #9608. I'm pretty sure there is sth I'm still missing though....

Benchmark with the following script
```julia
#!/usr/bin/julia

function f(a::Any, b::Any)
    global c = a + b * 2
end

function f(a::Integer, b::Integer)
    global c = a + b * 3
end

function f(a::Int, b::Int)
    global c = a + b * 4
end

macro timing(ex)
    quote
        println($(Expr(:quote, ex)))
        gc()
        @time for i in 1:10000000
            $(esc(ex))
        end
    end
end

function call_any()
    f(1.2, 3.4)
end

function call_integer()
    f(Int32(1), Int32(2))
end

function call_int()
    f(1, 2)
end

const f_any = (@which f(1.2, 3.4)).func
const f_integer = (@which f(Int32(1), Int32(2))).func
const f_int = (@which f(1, 2)).func

function meth_any()
    f_any(1.2, 3.4)
end

function meth_integer()
    f_integer(Int32(1), Int32(2))
end

function meth_int()
    f_int(1, 2)
end

function invoke_any()
    invoke(f, (Float64, Float64), 1.2, 3.4)
end

function invoke_integer()
    invoke(f, (Int32, Int32), Int32(1), Int32(2))
end

function invoke_int()
    invoke(f, (Int, Int), 1, 2)
end

function invoke_any_int()
    invoke(f, (Any, Any), 1, 2)
end

function invoke_integer_int()
    invoke(f, (Integer, Integer), 1, 2)
end

@timing call_any()
@timing call_integer()
@timing call_int()
println()
@timing meth_any()
@timing meth_integer()
@timing meth_int()
println()
@timing invoke_any()
@timing invoke_integer()
@timing invoke_int()
println()
@timing invoke_any_int()
@timing invoke_integer_int()
println()
```

Before,
```
call_any()
elapsed time: 0.150808637 seconds (160000000 bytes allocated, 37.06% gc time)
call_integer()
elapsed time: 0.041882309 seconds (0 bytes allocated)
call_int()
elapsed time: 0.041811296 seconds (0 bytes allocated)

meth_any()
elapsed time: 0.679259604 seconds (320001472 bytes allocated, 16.23% gc time)
meth_integer()
elapsed time: 0.4619528 seconds (5424 bytes allocated)
meth_int()
elapsed time: 0.452719865 seconds (1472 bytes allocated)

invoke_any()
elapsed time: 2.057187642 seconds (320002880 bytes allocated, 5.51% gc time)
invoke_integer()
elapsed time: 1.650398796 seconds (3136 bytes allocated)
invoke_int()
elapsed time: 1.255498639 seconds (2624 bytes allocated)

invoke_any_int()
elapsed time: 1.695137055 seconds (2112 bytes allocated)
invoke_integer_int()
elapsed time: 1.562752127 seconds (2112 bytes allocated)
```

After
```
call_any()
elapsed time: 0.140487996 seconds (160000000 bytes allocated, 35.90% gc time)
call_integer()
elapsed time: 0.04067296 seconds (0 bytes allocated)
call_int()
elapsed time: 0.041081744 seconds (0 bytes allocated)

meth_any()
elapsed time: 0.636057861 seconds (320001472 bytes allocated, 16.59% gc time)
meth_integer()
elapsed time: 0.364873117 seconds (5424 bytes allocated)
meth_int()
elapsed time: 0.389887527 seconds (1472 bytes allocated)

invoke_any()
elapsed time: 0.139580889 seconds (160000000 bytes allocated, 39.45% gc time)
invoke_integer()
elapsed time: 0.040794216 seconds (0 bytes allocated)
invoke_int()
elapsed time: 0.041527671 seconds (0 bytes allocated)

invoke_any_int()
elapsed time: 0.062740047 seconds (0 bytes allocated)
invoke_integer_int()
elapsed time: 0.063173429 seconds (0 bytes allocated)
```