@JeffBezanson 2015-01-06
Wow, this is impressive. Very good work. I'll review it in more detail later.

A related change that should help a lot, and is likely easier to implement than this, is to add a case to `emit_known_call` for `jl_f_invoke` in codegen.cpp. The code there should do the method lookup at compile time, and generate a direct call, as we do now for `jl_apply_generic`. This will eliminate most of the overhead for calls that can't be inlined.