@yuyichao 2015-01-07
I guess I'm still not very familiar with the order of these passes it still cannot inline invoke if it is wrapped....

```julia
yuyichao% cat invoke-inline.jl 
#!/usr/bin/julia -f

invoke_wrap(f::Function, ts::Tuple, args...) = invoke(f, ts, args...)

f(a, b) = a + b

g_invoke() = invoke(f, (Any, Any), 1, 2)
g_invoke_wrap() = invoke_wrap(f, (Any, Any), 1, 2)

println(@code_typed g_invoke())
println(@code_typed g_invoke_wrap())
yyc2:~/projects/explore/julia/invoke
yuyichao% ./invoke-inline.jl 
Any[:($(Expr(:lambda, Any[], Any[Any[:_var0,:_var1],Any[Any[:_var0,Int64,18],Any[:_var1,Int64,18]],Any[]], :(begin  # /home/yuyichao/projects/explore/julia/invoke/invoke-inline.jl, line 7:
        _var0 = 1
        _var1 = 2
        return (top(box))(Int64,(top(add_int))(_var0::Int64,_var1::Int64))
    end::Int64))))]
Any[:($(Expr(:lambda, Any[], Any[Any[:_var0,:_var1],Any[Any[:_var0,(Int64,Int64),0],Any[:_var1,Function,18]],Any[]], :(begin  # /home/yuyichao/projects/explore/julia/invoke/invoke-inline.jl, line 8:
        _var1 = f::ANY
        return invoke(_var1::F,(Any,Any),1,2)::ANY
    end::ANY))))]
```