@yuyichao 2015-01-07
Actually it seems that the inlining of function is not very smart in some cases either (when the function is assigned to a variable) even if the variable is a const

```julia
yuyichao% cat invoke-inline.jl
#!/usr/bin/julia -f

invoke_wrap(f::Function, ts::Tuple, args...) = invoke(f, ts, args...)

f(a, b) = a + b

g_invoke() = invoke(f, (Any, Any), 1, 2)
g_invoke_wrap() = invoke_wrap(f, (Any, Any), 1, 2)
function g_invoke2()
    const tmp_f = f
    invoke(tmp_f, (Any, Any), 1, 2)
end
function g_call2()
    const tmp_f = f
    tmp_f(1, 2)
end

println(@code_typed g_invoke())
println(@code_typed g_invoke_wrap())
println(@code_typed g_invoke2())
println(@code_typed g_call2())
yyc2:~/projects/explore/julia/invoke
yuyichao% ./invoke-inline.jl
Any[:($(Expr(:lambda, Any[], Any[Any[:_var0,:_var1],Any[Any[:_var0,Int64,18],Any[:_var1,Int64,18]],Any[]], :(begin  # /home/yuyichao/projects/explore/julia/invoke/invoke-inline.jl, line 7:
        _var0 = 1
        _var1 = 2
        return (top(box))(Int64,(top(add_int))(_var0::Int64,_var1::Int64))
    end::Int64))))]
Any[:($(Expr(:lambda, Any[], Any[Any[:_var0,:_var1],Any[Any[:_var0,(Int64,Int64),0],Any[:_var1,Function,18]],Any[]], :(begin  # /home/yuyichao/projects/explore/julia/invoke/invoke-inline.jl, line 8:
        _var1 = f::ANY
        return invoke(_var1::F,(Any,Any),1,2)::ANY
    end::ANY))))]
Any[:($(Expr(:lambda, Any[], Any[Any[:tmp_f],Any[Any[:tmp_f,Function,18]],Any[]], :(begin  # /home/yuyichao/projects/explore/julia/invoke/invoke-inline.jl, line 10:
        const tmp_f::ANY
        tmp_f = f::ANY # line 11:
        return invoke(tmp_f::F,(Any,Any),1,2)::ANY
    end::ANY))))]
Any[:($(Expr(:lambda, Any[], Any[Any[:tmp_f],Any[Any[:tmp_f,Function,18]],Any[]], :(begin  # /home/yuyichao/projects/explore/julia/invoke/invoke-inline.jl, line 14:
        const tmp_f::ANY
        tmp_f = f::ANY # line 15:
        return (tmp_f::F)(1,2)::ANY
    end::ANY))))]
```