@yuyichao 2015-01-08
@JeffBezanson Hopefully the last commit addresses some of your suggestions. There are some copy-paste's and I've also changed `expr_type` a little for supporint tuple types although it adds `TypeVar` at the same time....

```julia
julia> @noinline f(a, b) = a + b
f (generic function with 1 method)

julia> g() = invoke(f, (Any, Any), 1, 2)
g (generic function with 1 method)

julia> @code_typed g()
1-element Array{Any,1}:
 :($(Expr(:lambda, Any[], Any[Any[:_var0,:_var1],Any[Any[:_var0,Int64,18],Any[:_var1,Int64,18]],Any[]], :(begin  # none, line 1:
        return invoke(f,(Any,Any),1,2)::Int64
    end::Int64))))

julia> @code_llvm g();

define i64 @julia_g_42944() {
top:
  %0 = call i64 @julia_f_42945(i64 1, i64 2), !dbg !8
  ret i64 %0, !dbg !8
}
```