@yuyichao 2015-01-09
P.S. any objection if I make a `JL_GC_RETURN` macro that does `JL_GC_POP` and return? It is how 80-90% of `JL_GC_POP` are used.... (yes I counted...)