@vtjnash 2015-01-09
i thought most would be `value = compute(); JL_GC_POP; return value;`, which is slightly more difficult to combine into a macro (since C doesn't have gensym). however, you might be able to write that as a static inline function inside julia.h (although then you might have issues declaring the return type in a general enough way).

regardless, I wouldn't object.

> Another issue is that the lookup is done several times during compilation (3 times for a non-inlineable invoke). Not sure what is the best way to solve this...

I think method lookup does this too.