@yuyichao 2015-01-09
>
> i thought most would be value = compute(); JL_GC_POP; return value;,
>

I was thinking about just combinine the pop and return.

which is slightly more difficult to combine into a macro (since C doesn't
>

using a prefix in this case is probably good enough. It won't be worse than
the JL_TRY etc case. I'm not sure whether it is ok to use the GNU `typeof`
extension though. Otherwise it won't be possible to declare a varable with
the correct type in C.


> have gensym). however, you might be able to write that as a static inline
> function inside julia.h (although then you might have issues declaring the
> return type in a general enough way).
>

Would be possible in C++, although in that case there's much better way to
manage the GC stack.....


> regardless, I wouldn't object.
>
> Another issue is that the lookup is done several times during compilation
> (3 times for a non-inlineable invoke). Not sure what is the best way to
> solve this...
>
> I think method lookup does this too.
>

Yes it is. And probably not less than invoke, which is why I didn't bother
doing anything fancier for invoke either. However, IIRC, replacing the
really slow lookup in `invoke_tfunc` with the new C-api actually have a
noticeable improvement in compilation (type inference) time. I'm wondering
if it would be beneficial to improve this.....

> —
> Reply to this email directly or view it on GitHub
> <https://github.com/JuliaLang/julia/pull/9642#issuecomment-69279636>.
>