@tkelman 2015-01-09
cc @vtjnash our CI is failing assertions on mac and segfaulting on Windows. Any ideas how to fix?