@yuyichao 2015-01-09
@tkelman It might have sth to do with this PR but AFAIK it happens after I did a rebase on the current master.....

Also is there a reason that `expr_type` does not support tuple constant value? Adding it seems to break current code but I have no idea why....