@vtjnash 2015-01-09
We've seen those on master also. Best strategy is probably to replace that assert with jlbacktrace and dump a little more info about the environment