@vtjnash 2015-01-10
> Also is there a reason that expr_type does not support tuple constant value? Adding it seems to break current code but I have no idea why....

If I had to guess, it's probably because codegen assumes that if it comes across a type-tuple, it should be unboxed. but that isn't true of the current tuple implementation