@yuyichao 2015-01-11
Just noticed another issue. Since the inference of `_apply` does pass on the arguments, type infering of `_apply(invoke, ...)` fails although the call can now be inlined....